Refactor everything channel claims into folder in commons
