+++ customer-commons/src/main/java/no/unit/nva/customer/model/channelclaim/ChannelClaimDto.java
@@ -1,4 +1,4 @@
-package no.unit.nva.customer.model;
+package no.unit.nva.customer.model.channelclaim;
 
 import no.unit.nva.commons.json.JsonSerializable;
 import nva.commons.core.JacocoGenerated;
