+++ customer-commons/src/main/java/no/unit/nva/customer/model/channelclaim/ChannelConstraintDto.java
@@ -1,8 +1,9 @@
-package no.unit.nva.customer.model;
+package no.unit.nva.customer.model.channelclaim;
 
 import no.unit.nva.commons.json.JsonSerializable;
 
 import java.util.List;
+import no.unit.nva.customer.model.PublicationInstanceTypes;
 
 public record ChannelConstraintDto(ChannelConstraintPolicy publishingPolicy,
                                    ChannelConstraintPolicy editingPolicy,
