+++ customer-testing/src/main/java/no/unit/nva/customer/testing/CustomerDataGenerator.java
@@ -17,11 +17,11 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import no.unit.nva.customer.model.ApplicationDomain;
-import no.unit.nva.customer.model.ChannelClaimDao;
-import no.unit.nva.customer.model.ChannelClaimDto;
-import no.unit.nva.customer.model.ChannelConstraintDao;
-import no.unit.nva.customer.model.ChannelConstraintDto;
-import no.unit.nva.customer.model.ChannelConstraintPolicy;
+import no.unit.nva.customer.model.channelclaim.ChannelClaimDao;
+import no.unit.nva.customer.model.channelclaim.ChannelClaimDto;
+import no.unit.nva.customer.model.channelclaim.ChannelConstraintDao;
+import no.unit.nva.customer.model.channelclaim.ChannelConstraintDto;
+import no.unit.nva.customer.model.channelclaim.ChannelConstraintPolicy;
 import no.unit.nva.customer.model.CustomerDao;
 import no.unit.nva.customer.model.CustomerDao.RightsRetentionStrategyDao;
 import no.unit.nva.customer.model.CustomerDao.ServiceCenterDao;
