+++ customer/src/main/java/no/unit/nva/customer/create/ChannelClaimRequest.java
@@ -1,7 +1,7 @@
 package no.unit.nva.customer.create;
 
 import java.net.URI;
-import no.unit.nva.customer.model.ChannelClaimDto;
+import no.unit.nva.customer.model.channelclaim.ChannelClaimDto;
 
 public record ChannelClaimRequest(URI channel, ChannelConstraintRequest constraint) {
