+++ customer/src/main/java/no/unit/nva/customer/get/ListAllChannelClaimsHandler.java
@@ -3,13 +3,7 @@ package no.unit.nva.customer.get;
 import static java.net.HttpURLConnection.HTTP_OK;
 import static no.unit.nva.customer.Constants.defaultCustomerService;
 import com.amazonaws.services.lambda.runtime.Context;
-import java.util.AbstractMap.SimpleEntry;
-import java.util.stream.Collectors;
-import no.unit.nva.customer.get.response.ChannelClaimResponse;
 import no.unit.nva.customer.get.response.ChannelClaimsListResponse;
-import no.unit.nva.customer.get.response.CustomerResponse;
-import no.unit.nva.customer.model.ChannelClaimDto;
-import no.unit.nva.customer.model.CustomerDto;
 import no.unit.nva.customer.service.CustomerService;
 import nva.commons.apigateway.ApiGatewayHandler;
 import nva.commons.apigateway.RequestInfo;
