+++ customer/src/main/java/no/unit/nva/customer/get/response/ChannelClaimResponse.java
@@ -1,6 +1,6 @@
 package no.unit.nva.customer.get.response;
 
-import no.unit.nva.customer.model.ChannelClaimDto;
+import no.unit.nva.customer.model.channelclaim.ChannelClaimDto;
 
 public record ChannelClaimResponse(CustomerResponse claimedBy, ChannelClaimDto channelClaim) {
