+++ customer/src/test/java/no/unit/nva/customer/create/CreateChannelClaimHandlerTest.java
@@ -21,8 +21,8 @@ import java.net.HttpURLConnection;
 import java.util.Map;
 import java.util.UUID;
 import no.unit.nva.customer.exception.InputException;
-import no.unit.nva.customer.model.ChannelClaimDto;
-import no.unit.nva.customer.model.ChannelConstraintDto;
+import no.unit.nva.customer.model.channelclaim.ChannelClaimDto;
+import no.unit.nva.customer.model.channelclaim.ChannelConstraintDto;
 import no.unit.nva.customer.model.CustomerDto;
 import no.unit.nva.customer.service.CustomerService;
 import no.unit.nva.customer.service.impl.DynamoDBCustomerService;
